Refactor alt text generation module: update README, remove deprecated classes, and implement AltGeneratorTrait for improved functionality
