+++ composer.json
@@ -17,7 +17,8 @@
         "silverstripe/framework": "^5.0",
         "silverstripe/admin": "^2.0",
         "silverstripe/asset-admin": "^2.3",
-        "guzzlehttp/guzzle": "^7.9"
+        "guzzlehttp/guzzle": "^7.9",
+        "khalsa-jio/silverstripe-ai-nexus": "dev-main"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.6",
