+++ src/Client/Claude.php
@@ -2,48 +2,18 @@
 
 namespace KhalsaJio\AltGenerator\Client;
 
-use GuzzleHttp\RequestOptions;
+use KhalsaJio\AI\Nexus\Provider\Claude as ClaudeProvider;
+use KhalsaJio\AltGenerator\Client\AltGeneratorTrait;
 
-class Claude extends AbstractLLMClient
+class Claude extends ClaudeProvider
 {
-    /**
-     * API URL for Anthropic Claude API
-     * @var string
-     */
-    protected $apiUrl = 'https://api.anthropic.com/v1/messages';
+    use AltGeneratorTrait;
 
-    /**
-     * Get the default model to use
-     *
-     * @return string
-     */
-    protected function getDefaultModel(): string
+    private function __construct()
     {
-        return 'claude-3-haiku-20240307';
-    }
+        parent::__construct();
 
-    /**
-     * Get headers for API requests - Claude uses a different auth header
-     *
-     * @return array
-     */
-    protected function getRequestHeaders(): array
-    {
-        return [
-            'x-api-key' => $this->getApiKey(),
-            'anthropic-version' => '2023-06-01',
-            'Content-Type' => 'application/json',
-        ];
-    }
-
-    /**
-     * Get client name
-     *
-     * @return string
-     */
-    public static function getClientName(): string
-    {
-        return 'Claude';
+        $this->initAltGenerator();
     }
 
     /**
@@ -54,65 +24,29 @@ class Claude extends AbstractLLMClient
      */
     public function generateAltText($base_64_image)
     {
-        try {
-            $data = [
-                'model' => $this->getModel(),
-                'messages' => [
-                    [
-                        'role' => 'user',
-                        'content' => [
-                            [
-                                'type' => 'text',
-                                'text' => $this->getPrompt(),
-                            ],
-                            [
-                                'type' => 'image',
-                                'source' => [
-                                    'type' => 'base64',
-                                    'media_type' => 'image/webp',
-                                    'data' => $base_64_image
-                                ]
+        $data = [
+            'messages' => [
+                [
+                    'role' => 'user',
+                    'content' => [
+                        [
+                            'type' => 'text',
+                            'text' => $this->getPrompt(),
+                        ],
+                        [
+                            'type' => 'image',
+                            'source' => [
+                                'type' => 'base64',
+                                'media_type' => 'image/webp',
+                                'data' => $base_64_image
                             ]
                         ]
                     ]
-                ],
-                'max_tokens' => 300
-            ];
-
-            $response = $this->client->post('', [
-                RequestOptions::JSON => $data
-            ]);
-
-            $result = json_decode($response->getBody()->getContents(), true);
-
-            return $this->formatResponse($result);
-        } catch (\Exception $e) {
-            return $this->formatErrorResponse($e);
-        }
-    }
-
-    /**
-     * Extract alt text from response
-     *
-     * @param array $result
-     * @return string
-     */
-    protected function extractAltText($result): string
-    {
-        return trim($result['content'][0]['text'] ?? '');
-    }
-
-    /**
-     * Extract usage data from response
-     *
-     * @param array $result
-     * @return array
-     */
-    protected function extractUsageData($result): array
-    {
-        return [
-            'input_tokens' => $result['usage']['input_tokens'] ?? 0,
-            'output_tokens' => $result['usage']['output_tokens'] ?? 0
+                ]
+            ],
+            'max_tokens' => 300
         ];
+
+        return $this->chat($data, 'messages');
     }
 }
